Added a test script
This uses the python unittest module to bootstrap a small test
environment and then run two read tests (one sequential and a random).
This proved pretty useful as it found a bug when doing random reads. It
seems as FUSE does an initial read of 16384 bytes and so as long as we
keep reading within the first 16384 bytes we are OK, but if we seek
elsewhere and then read we get mismatches when comparing to the raw
file. Just looking in the rarfile source it appears as it doesn't
implement a seek() function at all, how is this supposed to work? ;)