Fugliest "solution" ever
This emulates offset / seek functionality by re-opening the file
(seek(0)) and then using read(offset) we get to the position we want,
after which we can do our read and thus return data from the correct
offset. Obviously this does not scale at all since we will be reading
enourmous amounts of data just to return a small chunk if offset is
anything but a tiny value.

Hardly know why I am commiting this, but I need this for a small project
(where performance doesn't matter) and so it goes to git.. expect a
better looking solution soonish.