Remove debug output & fix sequential read function
The sequential read function previously just read the first 1000 bytes
and used a seek(0) (though having previously just opened the file, this
shouldn't matter). Since we don't want to test seek() that was removed
and read() is now called without parameters to read the entire file and
not just the first 1000 bytes.