Remove offset test & reorganize code
The offset read is really not needed when we do the
random-read-from-start test. The actual tests were split into different
tests instead of having differenct functions for the tests and calling
them all from within the same test function (just view the diff). This
makes runtime for test longer as setUp() is called for each and every
test.