Use pythons logging class instead of homebrew
Ditch the dPrint (debugPrint) that was previously used to output debug
messages to syslog. Instead, pythons standard logging module has been
implemented and it is possible to send output to syslog as well as to
stdout if one wants to run pyarrfs in debugging mode. Configuration of
output and level is still done in the source file, which should
obviously be handled in a more elegant way but since we don't have any
option parsing support just yet, that's not possible.