Turned pyarrfs into a module
To ready pyarrfs for packaging using setup.py I've turned it into a
module. Much of this has been inspired by iotop, another application
written in python. To run pyarrfs non-installed, the file pyarrfs.py in
the root directory is provided while bin/pyarrfs is the one that is
going to be installed by setup.py