Implement own options to pyarr + fix logging
Finally got the hang of how fuse modifies optparse. It's not very pretty
IMHO but it works. So, we've added two options to specify root dir and
debugging, which is just debugging for the python part of PyarrFS and
not the FUSE part. This means we can now properly configure our logger
based on command line options instead of variables set inside our
script. Further, we integrate with the already existing FUSE options for
running in the foreground and -d for FUSE debug and do things
accordingly.