Seems other files are working now
That was a quick fix! I thought it would be lot more work, but I guess
not. TODO entry about it removed.

I've mangled the path in PyarrFile so it seems it is now able to open
files relative from a specified root (using --root). That is not working
for rar files yet thoug.