Refactor rar-path detection
Separate common functions for determining if a path is a rar archive, if
it is inside a rar archive and to split the path for what is inside and
what is outside the rar path if necessary. These functions are used in
multiple places and this should open up for the possibility to match
other .rar filenames more easily.