Wrap import in try/except and add usefull message
Non-standard python libs, ie those that do not come with Python by
default, are now wrapped in a try/except for their import and
appropriate messages are outputed if the user do not have the libs. For
fuse we check if we are running on a debian-like system and recommend
installing from that while for rarlib we point to git.spritelink.net
where the non-standard version required for PyarrFS resides.

If the user already has rarfile installed, it is quite likely that it
will silently fail as we do not check for it's version number.