Implement option to avoid compressed files
Compressed files are considerably slower to handle since rarfile cannot
use its DirectReader but has to use unrar (the binary) to first unpack
the files and the present them. Since this slowdown might not be
acceptable, here is an option to disable all rar files containing
compressed files. It will simply not present a .rar file as a directory
if it contains a compressed file.

It might be possible to do this on a more granular level, ie a rar
archive could contain both compressed and non-compressed file and we
could still present the rar archive as a directory, but I don't really
see how that would work.

Fixes #1.